+++ src/test/java/org/eclipse/tractusx/valueaddedservice/utils/PostgreSQLContextInitializer.java
@@ -1,0 +1,21 @@
+package org.eclipse.tractusx.valueaddedservice.utils;
+
+import org.springframework.boot.test.util.TestPropertyValues;
+import org.springframework.context.ApplicationContextInitializer;
+import org.springframework.context.ConfigurableApplicationContext;
+import org.testcontainers.containers.PostgreSQLContainer;
+
+public class PostgreSQLContextInitializer implements ApplicationContextInitializer<ConfigurableApplicationContext> {
+
+    public static final PostgreSQLContainer postgreSQLContainer = new PostgreSQLContainer("postgres:13.2");
+
+    @Override
+    public void initialize(ConfigurableApplicationContext applicationContext) {
+        postgreSQLContainer.start();
+        TestPropertyValues.of(
+                "spring.datasource.url=" + postgreSQLContainer.getJdbcUrl(),
+                "spring.datasource.username=" + postgreSQLContainer.getUsername(),
+                "spring.datasource.password=" + postgreSQLContainer.getPassword()
+        ).applyTo(applicationContext.getEnvironment());
+    }
+}